refactor(shared): rename camelCase files to kebab-case
configurableData.ts, isDeepEqual.ts, and settingsDataEntries.ts didn't
follow the project's kebab-case convention for TypeScript files; renamed
them and updated all import paths accordingly.
