+++ app/utils/index.ts
@@ -1,4 +1,4 @@
-import type { Selfmod } from "#shared/types/configurableData";
+import type { Selfmod } from "#shared/types/configurable-data";
 import type { Ref } from "vue";
 import { isEqual } from "ohash/utils";
