+++ shared/types/configurable-data.ts
