+++ shared/utils/index.ts
@@ -2,7 +2,7 @@ export * from "./discord";
 export * from "./bits";
 export * from "./comparators";
 export * from "./fetch-cache-config";
-export * from "./isDeepEqual";
-export * from "./settingsDataEntries";
+export * from "./is-deep-equal";
+export * from "./settings-data-entries";
 export * from "./times";
 export * from "./audit-field-metadata";
